refactor(server): simplify TLS server implementation and fix minor issues
Refactoring:
- Simplify tls_server.rs by removing unused imports and code
- Remove TlsIncoming, TlsStream and serve_https function
- Add helper methods to TlsManager for checking enabled features
- Update type imports to use local TlsConfig
- Add rustls-pki-types for PrivateKeyDer handling
- Simplify certificate and key loading logic

Minor Fixes:
- Fix type casting in handlers/mod.rs (unix.0 as u64 -> unix.0)
- Remove unused default_burst parameter in rate_limit.rs

+++ crates/server/src/handlers/mod.rs
@@ -260,7 +260,7 @@ impl ApiHandlers {
             .get_timestamp()
             .map(|ts| {
                 let unix = ts.to_unix();
-                unix.0 as u64 * 1000 + (unix.1 / 1_000_000) as u64
+                unix.0 * 1000 + (unix.1 / 1_000_000) as u64
             })
             .unwrap_or(0);
